Very questionable way to make phantom work with Capybara
The issue is that when Phantomjs makes the request to our server,
the server uses a different connection, which is outside of the
transaction that DatabaseCleaner has wrapped around the test.
So it cannot see the test data that we inserted, and possibly gets
put to sleep waiting for the transaction to end.

This change causes the server to use the same connection
as the tests, so it can see within the transaction.